Update app.py
Added ability for the code to stitch together parts where there is no dialogue, the ai likes to add "" (empty) dialogue and those would cause error if we did not add silence to those parts. Now, Pydub creates a silent bit and it is stitched in places where mp3 fails or is 0 bytes.